pyproject.toml
[project]
name = "gestao-vagas"
version = "0.1.0"
description = "Sistema interno de Requisicao de Pessoal com ingestao multimodal IA"
requires-python = ">=3.12"
dependencies = [
    "django>=5.0,<5.1",
    "django-environ>=0.11",
    "psycopg[binary]>=3.1",
    "django-redis>=5.4",
    "dramatiq[redis]>=1.17",
    "pgvector>=0.3",
    "django-stubs>=5.0",
    "mypy>=1.11",
    "ruff>=0.6",
    "pytest>=8.3",
    "pytest-django>=4.9",
    "factory-boy>=3.3",
    "structlog>=25.5.0",
    "django-waffle>=5.0.0",
    "pyotp>=2.9.0",
    "qrcode[pil]>=8.2",
    "cryptography>=46.0.7",
    "anthropic>=0.40",
    "openai>=1.50",
    "pydantic>=2.8",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=4.4",
    "freezegun>=1.5.5",
    "pytest-cov>=7.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["mypy_django_plugin.main"]

[[tool.mypy.overrides]]
module = ["apps.*.migrations.*"]
ignore_errors = true

[[tool.mypy.overrides]]
module = ["environ", "environ.*", "factory", "factory.*", "freezegun", "freezegun.*", "qrcode", "qrcode.*", "pyotp", "pyotp.*", "waffle", "waffle.*"]
ignore_missing_imports = true

# ADR-008: strict só em services/schemas/jobs. Testes ficam com type-check relaxado.
[[tool.mypy.overrides]]
module = ["apps.*.tests.*", "tests.*"]
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
disallow_untyped_decorators = false
warn_return_any = false
disable_error_code = ["type-arg", "no-untyped-call", "attr-defined", "arg-type", "assignment", "unused-ignore"]

# check_db/check_redis são management commands, não services críticos — relaxar strict
[[tool.mypy.overrides]]
module = ["apps.*.management.commands.*"]
disallow_untyped_defs = false

[tool.django-stubs]
django_settings_module = "config.settings.dev"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py", "tests.py"]
pythonpath = ["."]
